Crate islet_hes

Islet HES library.

Modules

security_lifecycle

Possible PSA lifecycle states (major):

Structs

AttestationMgr

Attestation manager implementing get_delegated_key and get_platform_token functinality.

BootMeasurement

Structure representing the boot measurement metadata and value.

BootMeasurementMetadata

Structure representing boot measurement metadata, as it is stored in emulated HW.

HWClaims

Hardware provisioned claims.

IsletHES

KeyMaterialData

Represents data required for key derivation (CPAK/DAK).

Measurement

Keeps measurement metadata and value.

MeasurementMetaData

Keeps measurement slot metadata.

MeasurementMgr

Responsible for storing all software components measurements and performing the read_measurement and extend_measurement functions.

Enums

AttestationError

Error kinds returned by AttestationMgr

ECCFamily

Supported ecc family types.

HashAlgo

Supported public dak hash algorithms.

IsletHESError

KeyBits

Supported ecc key bit size.

MeasurementError

Measurement error enumeration.

MeasurementType

Represents hash algorithm used for calculating measurement value.

Constants

MAX_HASH_VALUE_SIZE

MEASUREMENT_VALUE_MAX_SIZE

Maximal size based on the longest hash algorithm - sha512

MEASUREMENT_VALUE_MIN_SIZE

Minimal size based on the shortest hash algorithm - sha256

NUM_OF_MEASUREMENT_SLOTS

Maximum number of slots - based on the RSS implementation.

SIGNER_ID_MAX_SIZE

Maximal size based on the longest hash algorithm - sha512

SIGNER_ID_MIN_SIZE

Minimal size based on the shortest hash algorithm - sha256

SW_TYPE_MAX_SIZE

Set based on RSS imlementation.

VERSION_MAX_SIZE

Set based on RSS imlementation.

Traits

HWData

Interface for fetching hardware specific data:

Functions

calculate_public_key_hash

Calculate hash for given key with chosen HashAlgo

Type Aliases

HWAsymmetricKey

Represents a binary hardware asymmetric key. Currently using ECC Curve-P384 (384bit), might be a subject to change.

HWHash

Represents a hash value. Calculated by sha256, sha384 or sha512.

HWSWType

Represents a software type value.

HWSWVersion

Represents a software version value.

HWSymmetricKey

Represents a binary hardware 256bit symmetric key

SWType

SWVersion

SignerHash

ValueHash